Easy Coding Standard

+++ src/Parser/XmlParser.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace DTS\eBaySDK\Parser;
 
 /**
@@ -299,9 +300,9 @@ class XmlParser
     {
         switch ($meta->phpType) {
             case 'integer':
-                return (integer)$meta->strData;
+                return (int)$meta->strData;
             case 'double':
-                return (double)$meta->strData;
+                return (float)$meta->strData;
             case 'boolean':
                 return strtolower($meta->strData) === 'true';
             case 'DateTime':
@@ -326,11 +327,11 @@ class XmlParser
         } elseif (is_subclass_of($meta->phpObject, '\DTS\eBaySDK\Types\BooleanType', false)) {
             return strtolower($meta->strData) === 'true';
         } elseif (is_subclass_of($meta->phpObject, '\DTS\eBaySDK\Types\DecimalType', false)) {
-            return is_int(0 + $meta->strData) ? (integer)$meta->strData : (double)$meta->strData;
+            return is_int(0 + $meta->strData) ? (int)$meta->strData : (float)$meta->strData;
         } elseif (is_subclass_of($meta->phpObject, '\DTS\eBaySDK\Types\DoubleType', false)) {
-            return (double)$meta->strData;
+            return (float)$meta->strData;
         } elseif (is_subclass_of($meta->phpObject, '\DTS\eBaySDK\Types\IntegerType', false)) {
-            return (integer)$meta->strData;
+            return (int)$meta->strData;
         } elseif (is_subclass_of($meta->phpObject, '\DTS\eBaySDK\Types\StringType', false)) {
             return $meta->strData;
         } elseif (is_subclass_of($meta->phpObject, '\DTS\eBaySDK\Types\TokenType', false)) {
